Add an index.html file as a symlink
I'm adding a symlink from index.html to neat.html on the gh-pages
branch. My hope is that I can leave this branch alone and, because it's
a symlink, updates to the neat.html page will show up on gh-pages
automatically.

I'll still need to merge changes from master into gh-pages, but that's
all I might need to do on the gh-pages branch.