Remove div from center class
The center class only applied to div's before. Removing the div part
from the CSS makes it apply to any element you want to center in the
same way. This allows you to use p tags in a similar fashion. A positive
side-effect is that the CSS gets a few bytes smaller.