Remove the a tag from several other css classes
Previous the div tag was removed from the .center class. This saved a
few bytes and allowed the center class to be used on other elements,
such as a paragraph. Although I don't have a specific use for them, I
decided to do the same on the button and home classes. This reduces the
size of the CSS file at a minimum and offers a little more flexability.