Adds a `bordered` class for light on dark images
This adds a bordered class that can be used for images that don't stand out from the background. This was previously hard-coded into the neat.html file but this moves it into a class in the css. This also allows the border to work correctly in both light and dark mode.