Swap scheme order to light dark
The stylesheet is written with light values by default, and dark
values in a prefers-color-scheme media query. Therefore, this
should be color-scheme: light dark instead. As written, if no
preference is expressed, your styles will be light, but UA styles
will be dark.

Comment from Chris Morgan on Hacker News