Fix many css style issues outlined in HN comment
Chris Morgan left a long and detailed comment in the HN comment
board about some issues he saw in the css. I've gone through
those comments and made adjustments based on the feedback. I also
changed a few other tiny things along the way.

These are very likely breaking changes. Especially the point to
not automatically put all ol tags in two columns. I agree and
have moved that to a class. Because there are breaking changes,
because there are a lot here, and because I'm using Neat on a
lot of public projects now, I'm advancing the version number to
v1.0.0.