Update the release script to copy documentation index
I mainly modify the neat.html file right now when I'm working on
and testing changes. Because there is a docs directory I need to
copy the html file into the docs as index.html and then append
my tracking script to it. This change does both of this things.

Also fixes a minor issue from manually adding a version number
to the version.txt file.