Add button hover color change
Adds a slight color change on hover so that it's easier to see
that a button is a button. Also does a similar thing for inputs.
Adjusts dark mode colors for selected elements too. Fixes a typo
in a comment. Moves some styles to places that make more sense.
Adds some comments to identify why a choice was made.