Change wording back to "framework"
I've been torn on whether to call Neat a "CSS Framework" or not.
I've now thought more about this, referred to online resources,
and am deciding that, yes, it is a framework. I've added a section
describing my rationall for that. I will no longer listen to
pandentic comments about this.